fix(agent): wave7 — five surgical residuals; mutation diffs in log
- W7.1: remove post-drain ring re-plant; successor eats drain flush output
- W7.2: extract private runtime.ts; delete __testOnly export; zero-grep env name
- W7.4: exact off-nix / missing-BUILD_ID errors (mutations red on message swaps)
- W7.5: structured terminal outcomes (no message-prefix matching)
- W7.3: /tmp/w7-mutations.log sections include git diffs of each mutation
